+++ utils/download.js
@@ -1,25 +1,25 @@
 import fs from "fs-extra";
 import path from "path";
-import logger2 from "./logger2";
+import logger from "./logger";
 import MD5 from "js-md5";
 import axiosDirect from "axios";
 import {axiosProxied} from "./axiosProxied";
 
 async function download(url, useProxy = false) {
-    const axios = useProxy ? axiosDirect : axiosProxied;
+    const axios = useProxy ? axiosProxied : axiosDirect;
     let name = MD5(url);
-    logger2.info(["下载文件", url, "目标文件名", name].join(", "));
+    logger.info(["下载文件", url, "目标文件名", name].join(", "));
     
     //如果已经存在则直接返回路径
-    for (let ext of ["jpg", "mp4"]) {
-        let tmpdir = ext == "jpg" ? "./tmp/pic" : "./tmp/video";
-        if (!fs.existsSync("./tmp")) fs.mkdirSync("./tmp");
+    for (let ext of ["jpeg", "mp4"]) {
+        let tmpdir = ext == "jpeg" ? "tmp/pic" : "tmp/video";
+        if (!fs.existsSync("tmp")) fs.mkdirSync("tmp");
         if (!fs.existsSync(tmpdir)) fs.mkdirSync(tmpdir);
         
         let tmp = path.join(tmpdir, `${name}.${ext}`);
         if (fs.existsSync(tmp)) {
-            logger2.info("已存在文件: " + tmp);
-            return tmp;
+            logger.info("已存在文件: " + tmp);
+            return path.join(__dirname, "../", tmp);
         }
     }
     
@@ -27,9 +27,13 @@ async function download(url, useProxy = false) {
         url: url,
         method: "GET",
         responseType: "stream",
-        timeout: 3000,
+        timeout: 5000,
+        headers: {
+            "User-Agent": 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.93 Safari/537.36',
+            "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
+        }
     }).catch(err => {
-        logger2.warn("下载资源失败:", url, err);
+        logger.warn("下载资源失败:", url, err);
         return false;
     });
 
@@ -38,10 +42,10 @@ async function download(url, useProxy = false) {
         let mypath = "";
         let filename = `${name}.${mimeType}`;
 
-        if (["jpeg", "png", "gif"].some(t => t == mimeType)) mypath = path.join("./tmp/pic", filename);
-        else if (mimeType == "mp4") mypath = path.join("./tmp/video", filename);
+        if (["jpeg", "png", "gif"].some(t => t == mimeType)) mypath = path.join(__dirname, "../tmp/pic", filename);
+        else if (mimeType == "mp4") mypath = path.join(__dirname, "../tmp/video", filename);
         else {
-            logger2.warn("不支持下载的文件格式: " + mimeType, url);
+            logger.warn("不支持下载的文件格式: " + mimeType, url);
             return false;
         };
 
@@ -49,16 +53,16 @@ async function download(url, useProxy = false) {
         response.data.pipe(writer);
         return new Promise((resolve, reject) => {
             writer.on("finish", () => {
-                logger2.info("下载资源成功: " + filename);
+                logger.info("下载资源成功: " + filename);
                 resolve(mypath);
             });
             writer.on("error", err => {
-                logger2.warn("下载资源失败: " + url, err);
+                logger.warn("下载资源失败: " + url, err);
                 reject(err);
             });
         });
     } else {
-        logger2.warn("下载资源失败:", url, err);
+        logger.warn(["下载资源失败:", url].join(" "));
         return false;
     }
 }